requirements.txt
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.3
Babel==2.7.0
billiard==3.3.0.23
celery==3.1.15
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlog==4.1.0
ConfigArgParse==1.2.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
Django==3.0
django-cors-headers==3.2.1
django-filter==2.2.0
django-filter-stubs==0.1.1
django-stubs==1.5.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==1.2.0
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==1.5a3
geventhttpclient==1.4.2
geventhttpclient-wheels==1.3.1.dev2
greenlet==0.4.15
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.3
kombu==3.0.37
locustio==0.14.5
Markdown==3.1.1
MarkupSafe==1.1.1
msgpack==1.0.0
mypy==0.770
mypy-extensions==0.4.3
mysqlclient @ file:///C:/Users/dell/PycharmProjects/DeployDjango/django_app_docker/DjangoDev03/mysqlclient-1.4.6-cp38-cp38-win32.whl
psutil==5.7.0
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.1
requests==2.22.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
uritemplate==3.0.1
urllib3==1.25.7
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
